+++ .sampo/changesets/issue-375-migration-project-layout-helpers.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split migration project layout discovery and path helpers into focused runtime modules while preserving migration behavior.
